refactor: Voucher 도메인 모델에 맞게 할인값에 대한 변수 타입을 `int` -> `long` 으로 변경

+++ src/main/java/com/programmers/voucher/enumtype/VoucherType.java
@@ -52,7 +52,7 @@ public enum VoucherType {
                 });
     }
 
-    public void validateAmount(int amount) {
+    public void validateAmount(long amount) {
         voucherValidator.validateAmount(amount);
     }
